+++ Options.hpp
@@ -1,0 +1,20 @@
+// Options.hpp
+//
+#ifndef OPTIONS_HPP
+#define OPTIONS_HPP
+
+class Options
+{
+  private:
+
+  public:
+    bool continueLoop;
+    int userAction;
+    int userFont;
+    int userChoice;
+    Options();  // constructor
+    //int setFont(); // in graphics.hpp
+    //int setColor();  // in graphics.hpp
+};
+
+#endif
