+++ graphicstest.cc
@@ -1,0 +1,12 @@
+// Test driver for Graphics library module
+//
+#include <iostream>
+#include "graphics.hpp"
+#include "Options.hpp"
+
+using namespace std;
+
+int main()
+{
+  Graphics newGraphics;
+}
