+++ optiontest.cc
@@ -1,0 +1,13 @@
+// Test driver for Options library module
+//
+#include <iostream>
+#include "graphics.hpp"
+#include "Options.hpp"
+
+using namespace std;
+
+int main()
+{
+  Options newOptions;
+  //graphics newgraphics;
+}
