+++ testcase_menu_loops.txt
@@ -1,0 +1,4 @@
+1
+1
+2
+0
