+++ testcase_menu_walk.txt
@@ -1,0 +1,5 @@
+1
+1
+2
+0
+0
